Fix CustomEvent Implementation
Currently `new CustomEvent('!').constructor === CustomEvent` returns false, which is a hell of a bummer for anything else that is either polyfilling or feature detecting CustomEvent after this file is on the page.

Bonus: I have removed the redundant `in` check when `typeof` is all it's needed.

Please consider publishing this fix ASAP since it's conflicting with other polyfills.

Thank you.